AccessPoint: Changed HT to a String so that it would allow "Unknown" to be entered.
AirportScanner: Changed the HT output to a string.

Grid: Reformatted some things.

SampleProgram: Added the check for linux operating system and running the UbuntuScanner.

UbuntuScanner: Added UbuntuScanner which uses the iwlist command to scan for all available wireless networks and there signal strength. On ubuntu the program must be called using sudo. In the constructor it uses an enumerator method to list all available network devices and selects the first one that is a wlan.